SourceCode_and_Documentation/backend/requirements.txt
flask
python-dotenv
beautifulsoup4
pytz
flask-cors